refactor: ScoreManager파일 분리
